🚚 move ``sign_block_with_aggregation_bit_list`` to ``fork_choice.py``
